requirements.txt
flask
python-dotenv
requests
kucoin-python
ta
pandas
numpy